// src/types.ts
export interface CountriesData {
  name: {
    common: string;
    official: string;
    nativeName: { [abbr: string]: { official: string; common: string } };
  };
  tld: string[];
  cca2: string;
  cca3: string;
  ccn3: string;
  cioc: string;
  independent: boolean;
  status: string;
  unMember: boolean;
  currencies?: { [currency: string]: { name: string; symbol: string } };
  idd: { root: string; suffixes: string[] };
  capital: string[];
  altSpellings: string[];
  region: string;
  subregion: string;
  languages?: { [language: string]: string };
  translations: { [language: string]: { official: string; common: string } };
  langlng: number[];
  landlock: boolean;
  borders: string[];
  area: number;
  demonyms: { [demonym: string]: { f: string; m: string } };
  flag: string;
  maps: { googleMaps: string; openStreetMaps: string };
  population: number;
  gini: { [year: string]: number };
  fifa: string;
  timezones: string[];
  continents: string[];
  flags: { alt: string; png: string; svg: string };
  coatOfArms: { png: string; svg: string };
  startOfWeek: string;
  capitalInfo: { [latlng: string]: number[] };
  postalCode: { format: string; regex: string };
}